+++ src/MotherLanguage/MotherLanguage.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("../Language/Language.mjs").Language} Language */
+
+/**
+ * @typedef {Language} MotherLanguage
+ */