+++ src/Nationally/Nationally.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("../Label/Label.mjs").Label} Label */
+
+/**
+ * @typedef {{id: string, label: Label}} Nationally
+ */