+++ src/OriginPlace/OriginPlace.mjs
@@ -1,0 +1,5 @@
+/** @typedef {import("../Place/Place.mjs").Place} Place */
+
+/**
+ * @typedef {Place} OriginPlace
+ */